pip-requirements.txt
python-telegram-bot==11.1.0
urllib3~=1.26.4
requests~=2.25.1
pycurl~=7.43.0.5
dj-config-url~=0.1.1
psycopg2~=2.8.6
